add type support

+++ onnxruntime/core/providers/cuda/math/topk_impl.h
@@ -10,7 +10,7 @@ namespace onnxruntime {
 namespace cuda {
 
 template <typename T>
-Status TopKImpl(const T* input_x, T* output_v, int64_t* output_i, const int64_t* elem_nums, int64_t size, int64_t axis, int64_t K, int64_t largest, int64_t sorted);
+Status TopKImpl(const T* input_x, T* output_v, int64_t* output_i, const int64_t* elem_nums, size_t size, int64_t axis, int64_t K, int64_t largest, int64_t sorted);
 
 }  // namespace cuda
 }  // namespace onnxruntime
